requirements.txt
streamlit
pyautogen[gemini]


-e .